+++ PartyFpsTactics/Assets/_src/Scripts/WeaponsSystem/AttackColliders/ProjectileController.cs
@@ -73,7 +73,7 @@ namespace MrPink.WeaponsSystem
             PlaySound(flyAu);
         }
 
-        public override void Init(HealthController owner, DamageSource source, Transform shotHolder, ScoringActionType action = ScoringActionType.NULL, float offsetX = 0,float offsetY = 0, WeaponController weaponController = null, int damageScaler = 1)
+        public override void Init(HealthController owner, DamageSource source, Transform shotHolder, ScoringActionType action = ScoringActionType.NULL, float offsetX = 0,float offsetY = 0, WeaponController weaponController = null, float damageScaler = 1)
         {
             base.Init(owner, source, shotHolder, action);
             if (rb) rb.isKinematic = rbIsKinematicInit;
